+++ backend/src/main/java/com/morak/back/appointment/application/AppointmentService.java
@@ -7,12 +7,12 @@ import com.morak.back.appointment.domain.recommend.RankRecommendation;
 import com.morak.back.appointment.domain.recommend.RecommendationCells;
 import com.morak.back.appointment.exception.AppointmentAuthorizationException;
 import com.morak.back.appointment.exception.AppointmentNotFoundException;
-import com.morak.back.appointment.ui.dto.AppointmentAllResponse;
-import com.morak.back.appointment.ui.dto.AppointmentCreateRequest;
-import com.morak.back.appointment.ui.dto.AppointmentResponse;
-import com.morak.back.appointment.ui.dto.AppointmentStatusResponse;
-import com.morak.back.appointment.ui.dto.AvailableTimeRequest;
-import com.morak.back.appointment.ui.dto.RecommendationResponse;
+import com.morak.back.appointment.application.dto.AppointmentAllResponse;
+import com.morak.back.appointment.application.dto.AppointmentCreateRequest;
+import com.morak.back.appointment.application.dto.AppointmentResponse;
+import com.morak.back.appointment.application.dto.AppointmentStatusResponse;
+import com.morak.back.appointment.application.dto.AvailableTimeRequest;
+import com.morak.back.appointment.application.dto.RecommendationResponse;
 import com.morak.back.auth.domain.Member;
 import com.morak.back.core.application.AuthorizationService;
 import com.morak.back.core.domain.Code;
