+++ backend/src/main/java/com/morak/back/appointment/application/dto/AppointmentAllResponse.java
@@ -1,4 +1,4 @@
-package com.morak.back.appointment.ui.dto;
+package com.morak.back.appointment.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
