+++ backend/src/main/java/com/morak/back/auth/application/OAuthService.java
@@ -6,10 +6,10 @@ import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
 import com.morak.back.auth.exception.AuthenticationException;
 import com.morak.back.auth.exception.MemberNotFoundException;
-import com.morak.back.auth.ui.dto.ChangeNameRequest;
-import com.morak.back.auth.ui.dto.MemberResponse;
-import com.morak.back.auth.ui.dto.SigninRequest;
-import com.morak.back.auth.ui.dto.SigninResponse;
+import com.morak.back.auth.application.dto.ChangeNameRequest;
+import com.morak.back.auth.application.dto.MemberResponse;
+import com.morak.back.auth.application.dto.SigninRequest;
+import com.morak.back.auth.application.dto.SigninResponse;
 import com.morak.back.core.exception.CustomErrorCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
