+++ backend/src/main/java/com/morak/back/auth/application/dto/ChangeNameRequest.java
@@ -1,4 +1,4 @@
-package com.morak.back.auth.ui.dto;
+package com.morak.back.auth.application.dto;
 
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
