+++ backend/src/main/java/com/morak/back/core/application/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package com.morak.back.core.ui.dto;
+package com.morak.back.core.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
