+++ backend/src/main/java/com/morak/back/team/application/dto/InvitationJoinedResponse.java
@@ -1,4 +1,4 @@
-package com.morak.back.team.ui.dto;
+package com.morak.back.team.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
