+++ backend/src/main/java/com/morak/back/team/application/dto/TeamCreateRequest.java
@@ -1,4 +1,4 @@
-package com.morak.back.team.ui.dto;
+package com.morak.back.team.application.dto;
 
 import com.morak.back.core.domain.Code;
 import com.morak.back.core.domain.CodeGenerator;
