+++ backend/src/test/java/com/morak/back/appointment/application/AppointmentServiceTest.java
@@ -13,12 +13,12 @@ import com.morak.back.appointment.domain.AppointmentRepository;
 import com.morak.back.core.domain.SystemTime;
 import com.morak.back.appointment.exception.AppointmentAuthorizationException;
 import com.morak.back.appointment.exception.AppointmentNotFoundException;
-import com.morak.back.appointment.ui.dto.AppointmentAllResponse;
-import com.morak.back.appointment.ui.dto.AppointmentCreateRequest;
-import com.morak.back.appointment.ui.dto.AppointmentResponse;
-import com.morak.back.appointment.ui.dto.AppointmentStatusResponse;
-import com.morak.back.appointment.ui.dto.AvailableTimeRequest;
-import com.morak.back.appointment.ui.dto.RecommendationResponse;
+import com.morak.back.appointment.application.dto.AppointmentAllResponse;
+import com.morak.back.appointment.application.dto.AppointmentCreateRequest;
+import com.morak.back.appointment.application.dto.AppointmentResponse;
+import com.morak.back.appointment.application.dto.AppointmentStatusResponse;
+import com.morak.back.appointment.application.dto.AvailableTimeRequest;
+import com.morak.back.appointment.application.dto.RecommendationResponse;
 import com.morak.back.core.domain.Code;
 import com.morak.back.core.domain.CodeGenerator;
 import com.morak.back.core.domain.RandomCodeGenerator;
