+++ backend/src/test/java/com/morak/back/auth/acceptance/OAuthAcceptanceTest.java
@@ -8,9 +8,9 @@ import com.morak.back.AcceptanceTest;
 import com.morak.back.SimpleRestAssured;
 import com.morak.back.auth.application.FakeOAuthClient;
 import com.morak.back.auth.application.TokenProvider;
-import com.morak.back.auth.ui.dto.ChangeNameRequest;
-import com.morak.back.auth.ui.dto.MemberResponse;
-import com.morak.back.auth.ui.dto.SigninRequest;
+import com.morak.back.auth.application.dto.ChangeNameRequest;
+import com.morak.back.auth.application.dto.MemberResponse;
+import com.morak.back.auth.application.dto.SigninRequest;
 import com.morak.back.core.exception.CustomErrorCode;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
