+++ backend/src/test/java/com/morak/back/auth/application/OAuthServiceTest.java
@@ -5,10 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.morak.back.auth.application.dto.OAuthMemberInfoResponse;
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
-import com.morak.back.auth.ui.dto.ChangeNameRequest;
-import com.morak.back.auth.ui.dto.MemberResponse;
-import com.morak.back.auth.ui.dto.SigninRequest;
-import com.morak.back.auth.ui.dto.SigninResponse;
+import com.morak.back.auth.application.dto.ChangeNameRequest;
+import com.morak.back.auth.application.dto.MemberResponse;
+import com.morak.back.auth.application.dto.SigninRequest;
+import com.morak.back.auth.application.dto.SigninResponse;
 import com.morak.back.support.ServiceTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
