+++ backend/src/test/java/com/morak/back/notification/application/NotificationEventTest.java
@@ -7,8 +7,8 @@ import com.morak.back.appointment.domain.Appointment;
 import com.morak.back.appointment.domain.AppointmentEvent;
 import com.morak.back.appointment.domain.AppointmentRepository;
 import com.morak.back.core.domain.SystemTime;
-import com.morak.back.appointment.ui.dto.AppointmentCreateRequest;
-import com.morak.back.appointment.ui.dto.AppointmentResponse;
+import com.morak.back.appointment.application.dto.AppointmentCreateRequest;
+import com.morak.back.appointment.application.dto.AppointmentResponse;
 import com.morak.back.poll.application.PollService;
 import com.morak.back.poll.application.dto.PollCreateRequest;
 import com.morak.back.poll.application.dto.PollResponse;
