+++ backend/src/test/java/com/morak/back/team/acceptance/TeamAcceptanceTest.java
@@ -12,11 +12,11 @@ import com.morak.back.AcceptanceTest;
 import com.morak.back.AuthSupporter;
 import com.morak.back.SimpleRestAssured;
 import com.morak.back.auth.application.TokenProvider;
-import com.morak.back.auth.ui.dto.MemberResponse;
+import com.morak.back.auth.application.dto.MemberResponse;
 import com.morak.back.core.exception.CustomErrorCode;
-import com.morak.back.team.ui.dto.InvitationJoinedResponse;
-import com.morak.back.team.ui.dto.TeamCreateRequest;
-import com.morak.back.team.ui.dto.TeamResponse;
+import com.morak.back.team.application.dto.InvitationJoinedResponse;
+import com.morak.back.team.application.dto.TeamCreateRequest;
+import com.morak.back.team.application.dto.TeamResponse;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
